+++ HISTORY.md
@@ -3,13 +3,17 @@
 ## in development:
 * **New Features:**
 * **Notable Changes:**
-    * optimized color collision detection
 * **Notable Fixes:**
 * **Translation Updates:**
+
+## v5.3.8:
+* **Notable Change:**
+    * optimized color collision detection
 
 ### 2019-11-29
 * new dev version
 * objects, threads: optimized color collision detection
+* prepared release
 
 ## v5.3.7:
 * **Notable Fixes:**